Added weight and velocity initialization for gpu

+++ c++/gpu/include/data_reader.hpp
@@ -29,8 +29,8 @@ class data_reader
 		int size() { return m_images.size(); }
 		int rows() { return n_rows; }
 		int cols() { return n_cols; }
-
-  		std::vector<std::vector<double>> m_images;
+		
+  		std::vector<std::vector<float>> m_images;
 		std::vector<int> m_labels;
 };
 #endif
